+++ engine/sc_sim.cpp
@@ -1661,8 +1661,8 @@ void sim_t::analyze()
   {
     int last = ( int ) floor( iteration_timeline[ i ] );
     size_t num_buckets = divisor_timeline.size();
-    if ( 1 + last > num_buckets ) divisor_timeline.resize( 1 + last, 0 );
-    for ( size_t j=0; j <= last; j++ ) divisor_timeline[ j ] += 1;
+    if ( 1 + last > (int) num_buckets ) divisor_timeline.resize( 1 + last, 0 );
+    for ( int j=0; j <= last; j++ ) divisor_timeline[ j ] += 1;
   }
 
   // buff_t::analyze must be called before total_seconds is normalized via iteration count
